Moves invoice mapper into shared lib directory
