+++ src/lib/services/licences.js
@@ -3,7 +3,7 @@
 const repos = require('../connectors/repos');
 const licenceMapper = require('../mappers/licence');
 const invoiceLicenceMapper = require('../../modules/billing/mappers/invoice-licence');
-const invoiceMapper = require('../../modules/billing/mappers/invoice');
+const invoiceMapper = require('../mappers/invoice');
 const batchMapper = require('../../modules/billing/mappers/batch');
 const invoiceAccountsMapper = require('../mappers/invoice-account');
 const licenceVersionMapper = require('../mappers/licence-version');
