+++ src/modules/billing/mappers/index.js
@@ -7,7 +7,7 @@ exports.chargeVersion = require('../../../lib/mappers/charge-version');
 exports.chargeVersionYear = require('./charge-version-year');
 exports.company = require('../../../lib/mappers/company');
 exports.contact = require('../../../lib/mappers/contact');
-exports.invoice = require('./invoice');
+exports.invoice = require('../../../lib/mappers/invoice');
 exports.invoiceAccount = require('../../../lib/mappers/invoice-account');
 exports.invoiceAccountAddress = require('../../../lib/mappers/invoice-account-address');
 exports.invoiceLicence = require('./invoice-licence');
